requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
motor==3.1.1
pydantic==1.10.11
python-dotenv==1.0.1
openai==1.8.5
passlib[bcrypt]==1.7.5
pyjwt==2.8.0
